requirements.txt
python-telegram-bot==21.5
supabase==2.7.4
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.10.5
